requirements.txt
atomicwrites==1.2.1
attrs==18.2.0
boto==2.49.0
boto3==1.9.16
botocore==1.12.16
bz2file==0.98
certifi==2018.8.24
chardet==3.0.4
colorama==0.3.9
cycler==0.10.0
cymem==1.31.2
cytoolz==0.9.0.1
dataclasses==0.6
dill==0.2.8.2
Django==2.1.1
django-crispy-forms==1.7.2
django-rest-framework-mongoengine==3.3.1
djangorestframework==3.8.2
djongo==1.2.30
docutils==0.14
en-core-web-sm==2.0.0
funcy==1.11
future==0.16.0
gensim==3.6.0
idna==2.7
Jinja2==2.10
jmespath==0.9.3
joblib==0.12.5
kiwisolver==1.0.1
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==3.0.0
mongoengine==0.15.3
more-itertools==4.3.0
msgpack==0.5.6
msgpack-numpy==0.4.4.1
murmurhash==0.28.0
nltk==3.3
numexpr==2.6.8
numpy==1.15.2
pandas==0.23.4
plac==0.9.6
pluggy==0.7.1
preshed==1.0.1
py==1.6.0
pyLDAvis==2.1.2
pymongo==3.7.1
pyparsing==2.2.2
PyPDF2==1.26.0
pyreadline==2.1
pytest==3.8.2
python-dateutil==2.7.3
python-docx==0.8.7
pytz==2018.5
PyYAML==3.13
regex==2017.4.5
requests==2.19.1
s3transfer==0.1.13
scikit-learn==0.20.0
scipy==1.1.0
six==1.11.0
sklearn==0.0
smart-open==1.7.1
spacy==2.0.12
sqlparse==0.2.4
thinc==6.10.3
toolz==0.9.0
tqdm==4.26.0
ujson==1.35
urllib3==1.23
wrapt==1.10.11
